refactor: merge TestRunnerFactory to TestRunnerManager
